Add a pre-commit hook to use rhysd/actionlint
This is done through a pip wrapper package to pull down pre-built
binaries instead of relying on the Golang toolchain to build a binary
as would be done with the official repository's hook configuration. We
also add additional dependencies to enable actionlint to use pyflakes
to lint `run` steps using the `python` shell and to use shellcheck to
lint `run` steps using shellscript.